requirements.txt
face_recognition
alembic==1.4.0
astor==0.8.1
attrs==19.3.0
bandit==1.6.2
Click==7.0
cognitive-complexity==0.0.4
colorama==0.4.1
darglint==1.1.3
docutils==0.16
dominate==2.4.0
entrypoints==0.3
eradicate==1.0
Flask==1.1.1
Flask-Bootstrap==3.3.7.1
Flask-Login==0.5.0
Flask-Migrate==2.5.2
Flask-SQLAlchemy==2.4.1
future==0.18.2
gitdb2==3.0.2
GitPython==3.0.8
importlib-metadata==1.5.0
isort==4.3.21
itsdangerous==1.1.0
Jinja2==2.11.1
Mako==1.1.1
mando==0.6.4
MarkupSafe==1.1.1
mccabe==0.6.1
more-itertools==8.2.0
packaging==20.1
pbr==5.4.4
pep8-naming==0.9.1
pluggy==0.13.1
psycopg2==2.8.4
py==1.8.1
pycodestyle==2.5.0
pydocstyle==5.0.2
pyflakes==2.1.1
Pygments==2.5.2
pyparsing==2.4.6
pytest==5.3.5
python-dateutil==2.8.1
python-dotenv==0.11.0
python-editor==1.0.4
PyYAML==5.3
radon==4.1.0
restructuredtext-lint==1.3.0
six==1.14.0
smmap2==2.0.5
snowballstemmer==2.0.0
SQLAlchemy==1.3.13
stevedore==1.32.0
testfixtures==6.12.1
toml==0.10.0
typing-extensions==3.7.4.1
visitor==0.1.3
wcwidth==0.1.8
wemake-python-styleguide==0.13.4
Werkzeug==1.0.0
WTForms==2.2.1
WTForms-SQLAlchemy==0.1
zipp==2.2.0
